Adapt for new json struture and add link to original submission on assistant side
